chore(api): rename
