+++ packages/api/src/root.ts
@@ -1,7 +1,7 @@
 import { blankRouter } from "./router/blank";
 import { emailProviderRouter } from "./router/emailProvider";
 import { gmailRouter } from "./router/gmail";
-import { onboardingRouter } from "./router/onbaoarding";
+import { onboardingRouter } from "./router/onboarding";
 import { createTRPCRouter } from "./trpc";
 
 export const appRouter = createTRPCRouter({
